adding method to subscribe user to query
