+++ learning_resources_search/utils.py
@@ -1,7 +1,3 @@
-from learning_resources.hooks import get_plugin_manager
-from learning_resources_search.models import PercolateQuery
-
-
 def remove_child_queries(query):
     """
     Recursively removes 'has_child'/joins from a query
@@ -31,21 +27,3 @@ def remove_child_queries(query):
         else:
             return None
     return query
-
-
-def percolate_query_removed_actions(percolate_query: PercolateQuery):
-    """
-    Trigger plugins when a LearningResource is created or updated
-    """
-    pm = get_plugin_manager()
-    hook = pm.hook
-    hook.percolate_query_delete(percolate_query=percolate_query)
-
-
-def percolate_query_saved_actions(percolate_query: PercolateQuery):
-    """
-    Trigger plugins when a LearningResource is created or updated
-    """
-    pm = get_plugin_manager()
-    hook = pm.hook
-    hook.percolate_query_upserted(percolate_query=percolate_query)
